+++ src/main/java/com/backend/crmInmobiliario/DTO/entrada/ImpuestoLuzEntradaDto.java
@@ -9,9 +9,10 @@ import java.time.LocalDate;
 @Data
 @NoArgsConstructor
 public class ImpuestoLuzEntradaDto {
+
     private String descripcion;
     private String empresa;
-    private Double porcentaje;
+    private int porcentaje;
     private String numeroCliente;
     private String numeroMedidor;
     private Double montoAPagar;
